Calculate annual (and seasonal) cumulative flows

Calculates annual and seasonal total flows, volumetric or runoff yield flows, from a streamflow dataset. Calculates the statistics from all daily discharge values from all years, unless specified. For water year and seasonal data, the designated year is the year in which the year or season ends. Two-seasons and four-seasons per year are calculated, with each 6 and 3-month seasons starting with the first month of the year (Jan for calendar year, specified for water year). Each season is designated by the calendar or water year in which it occurs.

calc_annual_cumulative_stats(
  data,
  dates = Date,
  values = Value,
  groups = STATION_NUMBER,
  station_number,
  use_yield = FALSE,
  basin_area,
  water_year_start = 1,
  start_year,
  end_year,
  exclude_years,
  months = 1:12,
  include_seasons = FALSE,
  transpose = FALSE
)

Arguments

data
A data frame of daily data that contains columns of dates, flow values, and (optional) groups (e.g. station numbers). Leave blank if using station_number argument.

dates
Column in data that contains dates formatted YYYY-MM-DD. Only required if dates column name is not ' Date' (default). Leave blank if using station_number argument.

values
Column in data that contains numeric flow values, in units of cubic metres per second. Only required if values column name is not 'Value' (default). Leave blank if using station_number argument.

groups
Column in data that contains unique identifiers for different data sets, if applicable. Only required if groups column name is not 'STATION_NUMBER'. Function will automatically group by a column named 'STATION_NUMBER' if present. Remove the 'STATION_NUMBER' column beforehand to remove this grouping. Leave blank if using station_number argument.

station_number
A character string vector of seven digit Water Survey of Canada station numbers (e.g. "08NM116") of which to extract daily streamflow data from a HYDAT database. Requires tidyhydat package and a HYDAT database. Leave blank if using data argument.

use_yield
Logical value indicating whether to use yield runoff, in mm, instead of volumetric. Default FALSE.

basin_area
Upstream drainage basin area, in square kilometres, to apply to observations. Three options:

(1) Leave blank if groups is STATION_NUMBER with HYDAT station numbers to extract basin areas from HYDAT.

(2) A single numeric value to apply to all observations.

(3) List each basin area for each group/station in groups (can override HYDAT value if listed) as such c("08NM116" = 795, "08NM242" = 10). If group is not listed the HYDAT area will be applied if it exists, otherwise it will be NA.

water_year_start
Numeric value indicating the month of the start of the water year for analysis. Default 1.

start_year
Numeric value of the first year to consider for analysis. Leave blank to use the first year of the source data.

end_year
Numeric value of the last year to consider for analysis. Leave blank to use the last year of the source data.

exclude_years
Numeric vector of years to exclude from analysis. Leave blank to include all years.

months
Numeric vector of months to include in analysis (e.g.. 6:8 for Jun-Aug). Leave blank to summarize all months (default 1:12).

include_seasons
Logical value indication whether to include seasonal yields and total discharges. Default TRUE.

transpose
Logical value indicating if the results rows and columns are to be switched. Default FALSE.

Value

A tibble data frame with the following columns, ending with '_Volume_m3' or '_Yield_mm' based on selection:

Year

calendar or water year selected

Total_*

annual (or selected months) total flow, in m3 or mm

Default seasonal columns:
MMM-MMM_*

first of two season total flows, in m3 or mm

MMM-MMM_*

second of two season total flows, in m3 or mm

MMM-MMM_*

first of four season total flows, in m3 or mm

MMM-MMM_*

second of four season total flows, in m3 or mm

MMM-MMM_*

third of four season total flows, in m3 or mm

MMM-MMM_*

fourth of four season total flows, in m3 or mm

Transposing data creates a column of 'Statistics' and subsequent columns for each year selected.
Examples

# Run if HYDAT database has been downloaded (using tidyhydat::download_hydat())
if (file.exists(tidyhydat::hy_downloaded_db())) {

# Calculate annual total volumetric flow statistics
calc_annual_cumulative_stats(station_number = "08NM116")

# Calculate annual total yield statistics with default HYDAT basin area
calc_annual_cumulative_stats(station_number = "08NM116",
                             use_yield = TRUE)

# Calculate annual total yield statistics with a custom basin area
calc_annual_cumulative_stats(station_number = "08NM116",
                             use_yield = TRUE,
                             basin_area = 800)

}

#> Warning: One or more calculations included missing values and NA's were produced. Filter data for complete years or months, or use to ignore_missing = TRUE to ignore missing values.

#> Warning: One or more calculations included missing values and NA's were produced. Filter data for complete years or months, or use to ignore_missing = TRUE to ignore missing values.

#> Warning: One or more calculations included missing values and NA's were produced. Filter data for complete years or months, or use to ignore_missing = TRUE to ignore missing values.

#> # A tibble: 69 x 3
#>    STATION_NUMBER  Year Total_Yield_mm
#>    <chr>          <dbl>          <dbl>
#>  1 08NM116         1949             NA
#>  2 08NM116         1950             NA
#>  3 08NM116         1951             NA
#>  4 08NM116         1952             NA
#>  5 08NM116         1953             NA
#>  6 08NM116         1954             NA
#>  7 08NM116         1955             NA
#>  8 08NM116         1956             NA
#>  9 08NM116         1957             NA
#> 10 08NM116         1958             NA
#> # ... with 59 more rows