Calculate annual lowflows

Calculates annual n-day minimum values, and the day of year and date of occurrence of daily flow values from a streamflow dataset. Calculates the statistics from all daily discharge values from all years, unless specified.

calc_annual_lowflows(
  data,
  dates = Date,
  values = Value,
  groups = STATION_NUMBER,
  station_number,
  roll_days = c(1, 3, 7, 30),
  roll_align = "right",
  water_year_start = 1,
  start_year,
  end_year,
  exclude_years,
  months = 1:12,
  transpose = FALSE,
  ignore_missing = FALSE
)

Arguments

data
A data frame of daily data that contains columns of dates, flow values, and (optional) groups (e.g. station numbers). Leave blank if using station_number argument.

dates
Column in data that contains dates formatted YYYY-MM-DD. Only required if dates column name is not ' Date' (default). Leave blank if using station_number argument.

values
Column in data that contains numeric flow values, in units of cubic metres per second. Only required if values column name is not 'Value' (default). Leave blank if using station_number argument.

groups
Column in data that contains unique identifiers for different data sets, if applicable. Only required if groups column name is not 'STATION_NUMBER'. Function will automatically group by a column named 'STATION_NUMBER' if present. Remove the 'STATION_NUMBER' column beforehand to remove this grouping. Leave blank if using station_number argument.

station_number
A character string vector of seven digit Water Survey of Canada station numbers (e.g. "08NM116") of which to extract daily streamflow data from a HYDAT database. Requires tidyhydat package and a HYDAT database. Leave blank if using data argument.

roll_days
Numeric value of the number of days to apply a rolling mean. Default 1.

roll_align
Character string identifying the direction of the rolling mean from the specified date, either by the first ('left'), last ('right), or middle ('center') day of the rolling n-day group of observations. Default 'right'.

water_year_start
Numeric value indicating the month of the start of the water year for analysis. Default 1.

start_year
Numeric value of the first year to consider for analysis. Leave blank to use the first year of the source data.

end_year
Numeric value of the last year to consider for analysis. Leave blank to use the last year of the source data.

exclude_years
Numeric vector of years to exclude from analysis. Leave blank to include all years.

months
Numeric vector of months to include in analysis (e.g.. 6:8 for Jun-Aug). Leave blank to summarize all months (default 1:12).

transpose
Logical value indicating if the results rows and columns are to be switched. Default FALSE.

ignore_missing
Logical value indicating whether dates with missing values should be included in the calculation. If TRUE then a statistic will be calculated regardless of missing dates. If FALSE then only statistics from time periods with no missing dates will be returned. Default FALSE.

Value

A tibble data frame with the following columns:

Year

calendar or water year selected

Min_'n'_Day

annual minimum for each n-day rolling mean, direction of mean specified by roll_align

Min_'n'_Day_DoY

day of year for each annual minimum of n-day rolling mean

Min_'n'_Day_Date

date (YYYY-MM-DD) for each annual minimum of n-day rolling mean

Default columns:
Min_1_Day

annual 1-day mean minimum (roll_align = right)

Min_1_Day_DoY

day of year of annual 1-day mean minimum

Min_1_Day_Date

date (YYYY-MM-DD) of annual 1-day mean minimum

Min_3_Day

annual 3-day mean minimum (roll_align = right)

Min_3_Day_DoY

day of year of annual 3-day mean minimum

Min_3_Day_Date

date (YYYY-MM-DD) of annual 3-day mean minimum

Min_7_Day

annual 7-day mean minimum (roll_align = right)

Min_7_Day_DoY

day of year of annual 7-day mean minimum

Min_7_Day_Date

date (YYYY-MM-DD) of annual 7-day mean minimum

Min_30_Day

annual 30-day mean minimum (roll_align = right)

Min_30_Day_DoY

day of year of annual 30-day mean minimum

Min_30_Day_Date

date (YYYY-MM-DD) of annual 30-day mean minimum

Transposing data creates a column of 'Statistics' and subsequent columns for each year selected. 'Date' statistics not transposed.
Examples

# Run if HYDAT database has been downloaded (using tidyhydat::download_hydat())
if (file.exists(tidyhydat::hy_downloaded_db())) {

# Calculate annual 1, 3, 7, and 30-day (default) low flows with 
# default alignment ('right')
calc_annual_lowflows(station_number = "08NM116")

# Calculate custom 3 and 7-day annual low flows with 'center' alignment
calc_annual_lowflows(station_number = "08NM116",
                     roll_days = c(3,7),
                     roll_align = "center")

}

#> Warning: One or more calculations included missing values and NA's were produced. Filter data for complete years or months, or use to ignore_missing = TRUE to ignore missing values.

#> Warning: One or more calculations included missing values and NA's were produced. Filter data for complete years or months, or use to ignore_missing = TRUE to ignore missing values.

#> # A tibble: 69 x 8
#>    STATION_NUMBER  Year Min_3_Day Min_3_Day_DoY Min_3_Day_Date Min_7_Day
#>    <chr>          <dbl>     <dbl>         <dbl> <date>             <dbl>
#>  1 08NM116         1949        NA            NA NA                    NA
#>  2 08NM116         1950        NA            NA NA                    NA
#>  3 08NM116         1951        NA            NA NA                    NA
#>  4 08NM116         1952        NA            NA NA                    NA
#>  5 08NM116         1953        NA            NA NA                    NA
#>  6 08NM116         1954        NA            NA NA                    NA
#>  7 08NM116         1955        NA            NA NA                    NA
#>  8 08NM116         1956        NA            NA NA                    NA
#>  9 08NM116         1957        NA            NA NA                    NA
#> 10 08NM116         1958        NA            NA NA                    NA
#> # ... with 59 more rows, and 2 more variables: Min_7_Day_DoY <dbl>,
#> #   Min_7_Day_Date <date>