Calculate annual days above and below normal

Calculates the number of days per year outside of the 'normal' range (typically between 25 and 75th percentiles) for each day of the year. Upper and lower-range percentiles are calculated for each day of the year of from all years, and then each daily flow value for each year is compared. All days above or below the normal range are included. Calculates the statistics from all daily discharge values from all years, unless specified.

calc_annual_outside_normal(
  data,
  dates = Date,
  values = Value,
  groups = STATION_NUMBER,
  station_number,
  normal_percentiles = c(25, 75),
  roll_days = 1,
  roll_align = "right",
  water_year_start = 1,
  start_year,
  end_year,
  exclude_years,
  months = 1:12,
  transpose = FALSE
)

Arguments

data
A data frame of daily data that contains columns of dates, flow values, and (optional) groups (e.g. station numbers). Leave blank if using station_number argument.

dates
Column in data that contains dates formatted YYYY-MM-DD. Only required if dates column name is not ' Date' (default). Leave blank if using station_number argument.

values
Column in data that contains numeric flow values, in units of cubic metres per second. Only required if values column name is not 'Value' (default). Leave blank if using station_number argument.

groups
Column in data that contains unique identifiers for different data sets, if applicable. Only required if groups column name is not 'STATION_NUMBER'. Function will automatically group by a column named 'STATION_NUMBER' if present. Remove the 'STATION_NUMBER' column beforehand to remove this grouping. Leave blank if using station_number argument.

station_number
A character string vector of seven digit Water Survey of Canada station numbers (e.g. "08NM116") of which to extract daily streamflow data from a HYDAT database. Requires tidyhydat package and a HYDAT database. Leave blank if using data argument.

normal_percentiles
Numeric vector of two values, lower and upper percentiles, respectively indicating the limits of the normal range. Default c(25,75).

roll_days
Numeric value of the number of days to apply a rolling mean. Default 1.

roll_align
Character string identifying the direction of the rolling mean from the specified date, either by the first ('left'), last ('right), or middle ('center') day of the rolling n-day group of observations. Default 'right'.

water_year_start
Numeric value indicating the month of the start of the water year for analysis. Default 1.

start_year
Numeric value of the first year to consider for analysis. Leave blank to use the first year of the source data.

end_year
Numeric value of the last year to consider for analysis. Leave blank to use the last year of the source data.

exclude_years
Numeric vector of years to exclude from analysis. Leave blank to include all years.

months
Numeric vector of months to include in analysis (e.g.. 6:8 for Jun-Aug). Leave blank to summarize all months (default 1:12).

transpose
Logical value indicating if the results rows and columns are to be switched. Default FALSE.

Value

A tibble data frame with the following columns:

Year

calendar or water year selected

Days_Below_Normal

number of days per year below the daily normal (default 25th percentile)

Days_Above_Normal

number of days per year above the daily normal (default 75th percentile)

Days_Outside_Normal

number of days per year below and above the daily normal (default 25/75th percentile)

Transposing data creates a column of "Statistics" and subsequent columns for each year selected.
Examples

# Run if HYDAT database has been downloaded (using tidyhydat::download_hydat())
if (file.exists(tidyhydat::hy_downloaded_db())) {

# Calculate statistics with default limits of normal (25 and 75th percentiles)
calc_annual_outside_normal(station_number = "08NM116")

# Calculate statistics with custom limits of normal
calc_annual_outside_normal(station_number = "08NM116",
                           normal_percentiles = c(10,90))

}

#> Warning: One or more calculations included missing values and NA's were produced. Filter data for complete years or months, or use to ignore_missing = TRUE to ignore missing values.

#> Warning: One or more calculations included missing values and NA's were produced. Filter data for complete years or months, or use to ignore_missing = TRUE to ignore missing values.

#> # A tibble: 69 x 5
#>    STATION_NUMBER  Year Days_Below_Normal Days_Above_Normal Days_Outside_Normal
#>    <chr>          <dbl>             <int>             <int>               <int>
#>  1 08NM116         1949                NA                NA                  NA
#>  2 08NM116         1950                NA                NA                  NA
#>  3 08NM116         1951                NA                NA                  NA
#>  4 08NM116         1952                NA                NA                  NA
#>  5 08NM116         1953                NA                NA                  NA
#>  6 08NM116         1954                NA                NA                  NA
#>  7 08NM116         1955                NA                NA                  NA
#>  8 08NM116         1956                NA                NA                  NA
#>  9 08NM116         1957                NA                NA                  NA
#> 10 08NM116         1958                NA                NA                  NA
#> # ... with 59 more rows