Calculate annual summary statistics

Calculates means, medians, maximums, minimums, and percentiles for each year from all years of a daily streamflow data set. Calculates statistics from all values, unless specified. Returns a tibble with statistics.

Usage,

calc_annual_stats(
  data,
  dates = Date,
  values = Value,
  groups = STATION_NUMBER,
  station_number,
  roll_days = 1,
  roll_align = "right",
  percentiles = c(10, 90),
  water_year_start = 1,
  start_year,
  end_year,
  exclude_years,
  months = 1:12,
  transpose = FALSE,
  ignore_missing = FALSE,
  allowed_missing = ifelse(ignore_missing, 100, 0)
)

Arguments

data

Data frame of daily data that contains columns of dates, flow values, and (optional) groups (e.g. station numbers). Leave blank if using station_number argument.

dates

Name of column in data that contains dates formatted YYYY-MM-DD. Only required if dates column name is not 'Date' (default). Leave blank if using station_number argument.

values

Name of column in data that contains numeric flow values, in units of cubic metres per second. Only required if values column name is not 'Value' (default). Leave blank if using station_number argument.

groups

Name of column in data that contains unique identifiers for different data sets, if applicable. Only required if groups column name is not 'STATION_NUMBER'. Function will automatically group by a column named 'STATION_NUMBER' if present. Remove the 'STATION_NUMBER' column beforehand to remove this grouping. Leave blank if using station_number argument.

station_number

Character string vector of seven digit Water Survey of Canada station numbers (e.g. "08NM116") of which to extract daily streamflow data from a HYDAT database. Requires tidyhydat package and a HYDAT database. Leave blank if using data argument.

roll_days

Numeric value of the number of days to apply a rolling mean. Default 1.

roll_align

Character string identifying the direction of the rolling mean from the specified date, either by the first ('left'), last ('right'), or middle ('center') day of the rolling n-day group of observations. Default 'right'.

percentiles

Numeric vector of percentiles to calculate. Set to NA if none required. Default c(10,90).

water_year_start

Numeric value indicating the month (1 through 12) of the start of water year for analysis. Default 1.

start_year

Numeric value of the first year to consider for analysis. Leave blank to use the first year of the source data.

end_year

Numeric value of the last year to consider for analysis. Leave blank to use the last year of the source data.

exclude_years

Numeric vector of years to exclude from analysis. Leave blank to include all years.

months

Numeric vector of months to include in analysis (e.g. 6:8 for Jun-Aug). Leave blank to summarize all months (default 1:12).

transpose

Logical value indicating whether to transpose rows and columns of results. Default FALSE.

ignore_missing

Logical value indicating whether dates with missing values should be included in the calculation. If TRUE then a statistic will be calculated regardless of missing dates. If FALSE then only those statistics from time periods with no missing dates will be returned. Default FALSE.

allowed_missing

Numeric value between 0 and 100 indicating the percentage of missing dates allowed to be included to calculate a statistic (0 to 100 percent). If 'ignore_missing = FALSE' then it defaults to 0 (zero missing dates allowed), if 'ignore_missing = TRUE' then it defaults to 100 (any missing dates allowed); consistent with ignore_missing usage. Supersedes ignore_missing when used.

Value

A tibble data frame with the following columns:

Year

calendar or water year selected

Mean

annual mean of all daily flows for a given year

Median

annual median of all daily flows for a given year

Maximum

annual maximum of all daily flows for a given year

Minimum

annual minimum of all daily flows for a given year

P'n'

each annual n-th percentile selected of all daily flows

Default percentile columns:

P10

annual 10th percentile of all daily flows for a given year

P90

annual 90th percentile of all daily flows for a given year

Transposing data creates a column of "Statistics" and subsequent columns for each year selected.

Examples

# Run if HYDAT database has been downloaded (using tidyhydat::download_hydat())
if (file.exists(tidyhydat::hy_downloaded_db())) {

# Calculate annual statistics from a data frame using the data argument
flow_data <- tidyhydat::hy_daily_flows(station_number = "08NM116")
calc_annual_stats(data = flow_data)

# Calculate annual statistics using station_number argument
calc_annual_stats(station_number = "08NM116")

# Calculate annual statistics regardless if there 
# is missing data for a given year
calc_annual_stats(station_number = "08NM116",
                  ignore_missing = TRUE)
                  
# Calculate annual statistics for water years starting in October
calc_annual_stats(station_number = "08NM116",
                  water_year_start = 10)
                  
# Calculate annual statistics filtered for custom years
calc_annual_stats(station_number = "08NM116",
                  start_year = 1981,
                  end_year = 2010,
                  exclude_years = c(1991,1993:1995))
                  
# Calculate annual statistics for 7-day flows for July-September 
# months only, with 25 and 75th percentiles
calc_annual_stats(station_number = "08NM116",
                  roll_days = 7,
                  months = 7:9,
                  percentiles = c(25,75))
                  
}
#> Warning: One or more calculations included missing values and NA's were produced. If desired, filter data for complete years or months, or use the 'ignore_missing' or 'allowed_missing' arguments (if applicable) to ignore or allow some missing values.
#> Warning: One or more calculations included missing values and NA's were produced. If desired, filter data for complete years or months, or use the 'ignore_missing' or 'allowed_missing' arguments (if applicable) to ignore or allow some missing values.
#> Warning: One or more calculations included missing values and NA's were produced. If desired, filter data for complete years or months, or use the 'ignore_missing' or 'allowed_missing' arguments (if applicable) to ignore or allow some missing values.
#> Warning: One or more calculations included missing values and NA's were produced. If desired, filter data for complete years or months, or use the 'ignore_missing' or 'allowed_missing' arguments (if applicable) to ignore or allow some missing values.
#> # A tibble: 69 x 8
#>    STATION_NUMBER  Year  Mean Median Maximum Minimum   P25   P75
#>    <chr>          <dbl> <dbl>  <dbl>   <dbl>   <dbl> <dbl> <dbl>
#>  1 08NM116         1949  1.52   1.32    4.29   0.838 1.09   1.64
#>  2 08NM116         1950  2.71   1.36   16.6    0.714 0.860  2.10
#>  3 08NM116         1951  2.21   1.23    8.80   0.629 0.758  2.01
#>  4 08NM116         1952  2.85   1.93   12.9    1.11  1.32   2.56
#>  5 08NM116         1953  4.77   2.24   24.1    0.821 1.32   7.29
#>  6 08NM116         1954  7.65   4.34   30.5    1.98  3.14   8.88
#>  7 08NM116         1955  4.82   2.35   19.9    0.396 1.21   4.97
#>  8 08NM116         1956  2.25   1.29    8.21   0.784 0.998  2.33
#>  9 08NM116         1957  2.36   1.92    6.23   0.916 1.29   2.84
#> 10 08NM116         1958  1.71   1.23    5.51   0.578 0.880  2.04
#> # ... with 59 more rows