Calculate the percentile rank of a flow value

Calculates the percentile, or percentile rank, of a discharge value compared to all flow values of a streamflow dataset. Looks up the value in the distribution (stats::ecdf() function) of all daily discharge values from all years, unless specified.

calc_flow_percentile(
  data,
  dates = Date,
  values = Value,
  groups = STATION_NUMBER,
  station_number,
  roll_days = 1,
  roll_align = "right",
  flow_value,
  water_year_start = 1,
  start_year,
  end_year,
  exclude_years,
  complete_years = FALSE,
  months = 1:12
)

Arguments

data
A data frame of daily data that contains columns of dates, flow values, and (optional) groups (e.g. station numbers). Leave blank if using station_number argument.

dates
Column in data that contains dates formatted YYYY-MM-DD. Only required if dates column name is not ' Date' (default). Leave blank if using station_number argument.

values
Column in data that contains numeric flow values, in units of cubic metres per second. Only required if values column name is not 'Value' (default). Leave blank if using station_number argument.

groups
Column in data that contains unique identifiers for different data sets, if applicable. Only required if groups column name is not 'STATION_NUMBER'. Function will automatically group by a column named 'STATION_NUMBER' if present. Remove the 'STATION_NUMBER' column beforehand to remove this grouping. Leave blank if using station_number argument.

station_number
A character string vector of seven digit Water Survey of Canada station numbers (e.g. "08NM116") of which to extract daily streamflow data from a HYDAT database. Requires tidyhydat package and a HYDAT database. Leave blank if using data argument.

roll_days
Numeric value of the number of days to apply a rolling mean. Default 1.

roll_align
Character string identifying the direction of the rolling mean from the specified date, either by the first ('left'), last ('right), or middle ('center') day of the rolling n-day group of observations. Default 'right'.

flow_value
A numeric flow value of which to determine the percentile rank. Required.

water_year_start
Numeric value indicating the month of the start of the water year for analysis. Default 1.

start_year
Numeric value of the first year to consider for analysis. Leave blank to use the first year of the source data.

end_year
Numeric value of the last year to consider for analysis. Leave blank to use the last year of the source data.

exclude_years
Numeric vector of years to exclude from analysis. Leave blank to include all years.

complete_years
Logical values indicating whether to include only years with complete data in analysis. Default FALSE.

months
Numeric vector of months to include in analysis (e.g.. 6:8 for Jun-Aug). Leave blank to summarize all months (default 1:12).

Value

A tibble data frame, or a single numeric value if no station number provided, of the percentile rank of a given flow value.

Examples

# Run if HYDAT database has been downloaded (using tidyhydat::download_hydat())
if (file.exists(tidyhydat::hy_downloaded_db())) {

# Calculate the percentile rank of a 10-cms flow value from a full record
calc_flow_percentile(station_number = "08NM116",
                     flow_value = 10)

# Calculate the percentile rank of a 10-cms flow value from years with no missing data
calc_flow_percentile(station_number = "08NM116",
                     complete_years = TRUE,
                     flow_value = 10)

# Calculate the percentile rank of a 10-cms flow value for June from years with no missing data
calc_flow_percentile(station_number = "08NM116",
                     complete_years = TRUE,
                     months = 6,
                     flow_value = 10)

}

#> Warning: Calculation ignored missing values in data. Filter data to complete years or months if desired.

#> # A tibble: 1 x 2
#>   STATION_NUMBER Percentile
#>   <chr>               <dbl>
#> 1 08NM116              20.5