Fill dates of missing flow values with NA

Adds rows of dates with missing flow values to a streamflow dataset with daily flow values of NA. Missing dates will be filled in gaps between data and completely fill the first and last years (calendar or water year if selected).

fill_missing_dates(
  data,
  dates = Date,
  values = Value,
  groups = STATION_NUMBER,
  station_number,
  water_year_start = 1
)

Arguments

data
A data frame of daily data that contains columns of dates, flow values, and (optional) groups (e.g. station numbers). Leave blank if using station_number argument.

dates
Column in data that contains dates formatted YYYY-MM-DD. Only required if dates column name is not ' Date' (default). Leave blank if using station_number argument.

values
Column in data that contains numeric flow values, in units of cubic metres per second. Only required if values column name is not 'Value' (default). Leave blank if using station_number argument.

groups
Column in data that contains unique identifiers for different data sets, if applicable. Only required if groups column name is not 'STATION_NUMBER'. Function will automatically group by a column named 'STATION_NUMBER' if present. Remove the 'STATION_NUMBER' column beforehand to remove this grouping. Leave blank if using station_number argument.

station_number
A character string vector of seven digit Water Survey of Canada station numbers (e.g. "08NM116") of which to extract daily streamflow data from a HYDAT database. Requires tidyhydat package and a HYDAT database. Leave blank if using data argument.

water_year_start
Numeric value indicating the month of the start of the water year for analysis. Default 1.

Value

A tibble data frame of the source data with additional rows of filled values of missing dates.

Examples

# Run if HYDAT database has been downloaded (using tidyhydat::download_hydat())
if (file.exists(tidyhydat::hy_downloaded_db())) {

# Fill missing dates with NA using calendar years
fill_missing_dates(station_number = "08NM116")

# Fill missing dates with NA using water years starting in August
fill_missing_dates(station_number = "08NM116",
                   water_year_start = 8)

}

#> # A tibble: 25,567 x 5
#>    STATION_NUMBER Date       Parameter Value Symbol
#>    <chr>          <date>     <chr>     <dbl> <chr> 
#>  1 08NM116        1948-08-01 Flow         NA NA    
#>  2 08NM116        1948-08-02 Flow         NA NA    
#>  3 08NM116        1948-08-03 Flow         NA NA    
#>  4 08NM116        1948-08-04 Flow         NA NA    
#>  5 08NM116        1948-08-05 Flow         NA NA    
#>  6 08NM116        1948-08-06 Flow         NA NA    
#>  7 08NM116        1948-08-07 Flow         NA NA    
#>  8 08NM116        1948-08-08 Flow         NA NA    
#>  9 08NM116        1948-08-09 Flow         NA NA    
#> 10 08NM116        1948-08-10 Flow         NA NA    
#> # ... with 25,557 more rows