Calculate annual summary and missing data statistics for screening data

Calculates means, medians, maximums, minimums, standard deviations of annual flows and data availability and missing data statistics, and symbol counts (if column exists) for each year and month of each year. Calculates the statistics from all daily discharge values from all years, unless specified. Returns a tibble with statistics.

screen_flow_data(
  data,
  dates = Date,
  values = Value,
  groups = STATION_NUMBER,
  symbols = "Symbol",
  station_number,
  roll_days = 1,
  roll_align = "right",
  water_year_start = 1,
  start_year,
  end_year,
  months = 1:12,
  transpose = FALSE,
  include_symbols = TRUE
)

Arguments

data

Data frame of daily data that contains columns of dates, flow values, and (optional) groups (e.g. station numbers). Leave blank or set to NULL if using station_number argument.

dates

Name of column in data that contains dates formatted YYYY-MM-DD. Only required if dates column name is not 'Date' (default). Leave blank or set to NULL if using station_number argument.

values

Name of column in data that contains numeric flow values, in units of cubic metres per second. Only required if values column name is not 'Value' (default). Leave blank if using station_number argument.

groups

Name of column in data that contains unique identifiers for different data sets, if applicable. Only required if groups column name is not 'STATION_NUMBER'. Function will automatically group by a column named 'STATION_NUMBER' if present. Remove the 'STATION_NUMBER' column beforehand to remove this grouping. Leave blank if using station_number argument.

symbols

Name of column in data that contains symbols. Only required if symbols column name is not 'Symbol' (default). Leave blank or set to NULL if using station_number argument.

station_number

Character string vector of seven digit Water Survey of Canada station numbers (e.g. "08NM116") of which to extract daily streamflow data from a HYDAT database. Requires tidyhydat package and a HYDAT database. Leave blank if using data argument.

roll_days

Numeric value of the number of days to apply a rolling mean. Default 1.

roll_align

Character string identifying the direction of the rolling mean from the specified date, either by the first ('left'), last ('right'), or middle ('center') day of the rolling n-day group of observations. Default 'right'.

water_year_start

Numeric value indicating the month (1 through 12) of the start of water year for analysis. Default 1.

start_year

Numeric value of the first year to consider for analysis. Leave blank or set well before start date (i.e. 1800) to use from the first year of the source data.

end_year

Numeric value of the last year to consider for analysis. Leave blank or set well after end date (i.e. 2100) to use up to the last year of the source data.

months

Numeric vector of months to include in analysis. For example, 3 for March, 6:8 for Jun-Aug or c(10:12,1) for first four months (Oct-Jan) when water_year_start = 10 (Oct). Default summarizes all months (1:12).

transpose

Logical value indicating whether to transpose rows and columns of results. Default FALSE.

include_symbols

Logical. Include columns of counts of symbol categories from the symbols column.

Value

A tibble data frame with the following columns:

Year

calendar or water year selected

n_days

number of days per year

n_Q

number of days per year with flow data

n_missing_Q

number of days per year with no flow data

No_Symbol

number of days with no symbol category, if include_symbol=TRUE

x_Symbol

number of days with a specific symbol category (x) from symbols column, if include_symbol=TRUE

Maximum

annual maximum of all daily flows for a given year

Mean

annual mean of all daily flows for a given year

Median

annual median of all daily flows for a given year

StandardDeviation

annual 1 standard deviation of all daily flows for a given year

and the following monthly missing columns (order will depend on water_year_month):

Jan_missing_Q

number of Jan days per year with no flow data

Feb_missing_Q

number of Feb days per year with no flow data

Mar_missing_Q

number of Mar days per year with no flow data

Apr_missing_Q

number of Apr days per year with no flow data

May_missing_Q

number of May days per year with no flow data

Jun_missing_Q

number of Jun days per year with no flow data

Jul_missing_Q

number of Jul days per year with no flow data

Aug_missing_Q

number of Aug days per year with no flow data

Sep_missing_Q

number of Sep days per year with no flow data

Oct_missing_Q

number of Oct days per year with no flow data

Nov_missing_Q

number of Nov days per year with no flow data

Dec_missing_Q

number of Dec days per year with no flow data

Transposing data creates a column of "Statistics" and subsequent columns for each year selected.

Examples

# Run if HYDAT database has been downloaded (using tidyhydat::download_hydat())
if (file.exists(tidyhydat::hy_downloaded_db())) {

# Calculate screening statistics using data frame and data argument with defaults
flow_data <- tidyhydat::hy_daily_flows(station_number = "08NM116")
screen_flow_data(data = flow_data)

# Calculate screening statistics using station_number argument with defaults
screen_flow_data(station_number = "08NM116")
                  
# Calculate screening statistics for water years starting in October
screen_flow_data(station_number = "08NM116",
                 water_year_start = 9)
                  
# Calculate screening statistics for 7-day flows for July-September months only
screen_flow_data(station_number = "08NM116",
                 roll_days = 7,
                 months = 7:9)
                 
}
#> # A tibble: 73 × 16
#>    STATION_NUMBER  Year n_days   n_Q n_missing_Q E_Symbol No_Symbol A_Symbol
#>    <chr>          <dbl>  <int> <int>       <int>    <int>     <int>    <int>
#>  1 08NM116         1949     92    92           0        5        87        0
#>  2 08NM116         1950     92    92           0        0        92        0
#>  3 08NM116         1951     92    92           0        0        92        0
#>  4 08NM116         1952     92    92           0        0        92        0
#>  5 08NM116         1953     92    92           0        0        92        0
#>  6 08NM116         1954     92    92           0        0        92        0
#>  7 08NM116         1955     92    92           0        0        92        0
#>  8 08NM116         1956     92    92           0        0        92        0
#>  9 08NM116         1957     92    92           0        0        92        0
#> 10 08NM116         1958     92    92           0        0        92        0
#> # ℹ 63 more rows
#> # ℹ 8 more variables: Minimum <dbl>, Maximum <dbl>, Mean <dbl>, Median <dbl>,
#> #   StandardDeviation <dbl>, Jul_missing_Q <int>, Aug_missing_Q <int>,
#> #   Sep_missing_Q <int>