Calculate all fasstr annual statistics

Calculates all annual statistics of daily flow values from a streamflow dataset from all annual fasstr functions. Calculates the statistics from all daily discharge values from all years, unless specified. Data is ideally long-term and continuous with minimal missing/seasonal data as annual statistics are calculated. Data calculated using the folling functions:

calc_annual_stats()

calc_annual_lowflows()

calc_annual_cumulative_stats()

calc_annual_flow_timing()

calc_monthly_stats()

calc_annual_outside_normal()

calc_all_annual_stats(
  data,
  dates = Date,
  values = Value,
  groups = STATION_NUMBER,
  station_number,
  basin_area,
  water_year_start = 1,
  start_year,
  end_year,
  exclude_years,
  annual_percentiles = c(10, 90),
  monthly_percentiles = c(10, 20),
  stats_days = 1,
  stats_align = "right",
  lowflow_days = c(1, 3, 7, 30),
  lowflow_align = "right",
  timing_percent = c(25, 33, 50, 75),
  normal_percentiles = c(25, 75),
  transpose = FALSE,
  ignore_missing = FALSE
)

Arguments

data
A data frame of daily data that contains columns of dates, flow values, and (optional) groups (e.g. station numbers). Leave blank if using station_number argument.

dates
Column in data that contains dates formatted YYYY-MM-DD. Only required if dates column name is not ' Date' (default). Leave blank if using station_number argument.

values
Column in data that contains numeric flow values, in units of cubic metres per second. Only required if values column name is not 'Value' (default). Leave blank if using station_number argument.

groups
Column in data that contains unique identifiers for different data sets, if applicable. Only required if groups column name is not 'STATION_NUMBER'. Function will automatically group by a column named 'STATION_NUMBER' if present. Remove the 'STATION_NUMBER' column beforehand to remove this grouping. Leave blank if using station_number argument.

station_number
A character string vector of seven digit Water Survey of Canada station numbers (e.g. "08NM116") of which to extract daily streamflow data from a HYDAT database. Requires tidyhydat package and a HYDAT database. Leave blank if using data argument.

basin_area
Upstream drainage basin area, in square kilometres, to apply to observations. Three options:

(1) Leave blank if groups is STATION_NUMBER with HYDAT station numbers to extract basin areas from HYDAT.

(2) A single numeric value to apply to all observations.

(3) List each basin area for each group/station in groups (can override HYDAT value if listed) as such c("08NM116" = 795, "08NM242" = 10). If group is not listed the HYDAT area will be applied if it exists, otherwise it will be NA.

water_year_start
Numeric value indicating the month of the start of the water year for analysis. Default 1.

start_year
Numeric value of the first year to consider for analysis. Leave blank to use the first year of the source data.

end_year
Numeric value of the last year to consider for analysis. Leave blank to use the last year of the source data.

exclude_years
Numeric vector of years to exclude from analysis. Leave blank to include all years.

annual_percentiles
Numeric vector of percentiles to calculate annually. Set to NA if none required. Used for calc_annual_stats() function. Default c(10,90).

monthly_percentiles
Numeric vector of percentiles to calculate monthly for each year. Set to NA if none required. Used for calc_monthly_stats() function. Default c(10,20).

stats_days
Numeric vector of the number of days to apply a rolling mean on basic stats. Default c(1). Used for calc_annual_stats() and calc_monthly_stats() functions.

stats_align
Character string identifying the direction of the rolling mean on basic stats from the specified date, either by the first ('left'), last ('right), or middle ('center') day of the rolling n-day group of observations. Default 'right'. Used for calc_annual_stats(), calc_monthly_stats(), and calc_annual_outside_normal() functions.

lowflow_days
Numeric vector of the number of days to apply a rolling mean on lowflow stats. Default c(1,3,7,30). Used for calc_lowflow_stats() function.

lowflow_align
Character string identifying the direction of the rolling mean on lowflow stats from the specified date, either by the first ('left'), last ('right), or middle ('center') day of the rolling n-day group of observations. Default 'right'. Used for calc_lowflow_stats() function.

timing_percent
Numeric vector of percents of annual total flows to determine dates. Used for calc_annual_flow_timing() function. Default c(25,33.3,50,75).

normal_percentiles
Numeric vector of two values, lower and upper percentiles, respectively indicating the limits of the normal range. Default c(25,75).

transpose
Logical value indicating if the results rows and columns are to be switched. Default FALSE.

ignore_missing
Logical value indicating whether dates with missing values should be included in the calculation. If TRUE then a statistic will be calculated regardless of missing dates. If FALSE then only statistics from time periods with no missing dates will be returned. Default FALSE.

Value

A tibble data frame with column "Year" and then 107 (default) variables from the fasstr annual functions. See listed functions above for default variables. Transposing data creates a column of "Statistics" and subsequent columns for each year selected.

See also

Examples

# Run if HYDAT database has been downloaded (using tidyhydat::download_hydat())
if (file.exists(tidyhydat::hy_downloaded_db())) {

# Calculate all annual statistics from this package with default arguments
calc_all_annual_stats(station_number = "08NM116")

# Calculate all annual statistics from this package with default arguments 
# with some default arguments shown to customize metrics
calc_all_annual_stats(station_number = "08NM116",
                      annual_percentiles = c(10,90),
                      monthly_percentiles = c(10,20),
                      stats_days = 1,
                      stats_align = "right",
                      lowflow_days = c(1,3,7,30),
                      lowflow_align = "right",
                      timing_percent = c(25,33,50,75),
                      normal_percentiles = c(25,75))

}

#> Warning: One or more calculations included missing values and NA's were produced. Filter data for complete years or months, or use to ignore_missing = TRUE to ignore missing values.

#> Warning: One or more calculations included missing values and NA's were produced. Filter data for complete years or months, or use to ignore_missing = TRUE to ignore missing values.

#> # A tibble: 69 x 109
#>    STATION_NUMBER  Year Annual_Maximum Annual_Mean Annual_Median Annual_Minimum
#>    <chr>          <dbl>          <dbl>       <dbl>         <dbl>          <dbl>
#>  1 08NM116         1949             NA          NA            NA             NA
#>  2 08NM116         1950             NA          NA            NA             NA
#>  3 08NM116         1951             NA          NA            NA             NA
#>  4 08NM116         1952             NA          NA            NA             NA
#>  5 08NM116         1953             NA          NA            NA             NA
#>  6 08NM116         1954             NA          NA            NA             NA
#>  7 08NM116         1955             NA          NA            NA             NA
#>  8 08NM116         1956             NA          NA            NA             NA
#>  9 08NM116         1957             NA          NA            NA             NA
#> 10 08NM116         1958             NA          NA            NA             NA
#> # ... with 59 more rows, and 103 more variables: Annual_P10 <dbl>,
#> #   Annual_P90 <dbl>, Min_1_Day <dbl>, Min_1_Day_DoY <dbl>, Min_3_Day <dbl>,
#> #   Min_3_Day_DoY <dbl>, Min_7_Day <dbl>, Min_7_Day_DoY <dbl>,
#> #   Min_30_Day <dbl>, Min_30_Day_DoY <dbl>, Total_Volume_m3 <dbl>,
#> #   `Jan-Jun_Volume_m3` <dbl>, `Jul-Dec_Volume_m3` <dbl>,
#> #   `Jan-Mar_Volume_m3` <dbl>, `Apr-Jun_Volume_m3` <dbl>,
#> #   `Jul-Sep_Volume_m3` <dbl>, `Oct-Dec_Volume_m3` <dbl>, Total_Yield_mm <dbl>,
#> #   `Jan-Jun_Yield_mm` <dbl>, `Jul-Dec_Yield_mm` <dbl>,
#> #   `Jan-Mar_Yield_mm` <dbl>, `Apr-Jun_Yield_mm` <dbl>,
#> #   `Jul-Sep_Yield_mm` <dbl>, `Oct-Dec_Yield_mm` <dbl>, DoY_25pct_TotalQ <dbl>,
#> #   DoY_33pct_TotalQ <dbl>, DoY_50pct_TotalQ <dbl>, DoY_75pct_TotalQ <dbl>,
#> #   Days_Below_Normal <dbl>, Days_Above_Normal <dbl>,
#> #   Days_Outside_Normal <dbl>, Jan_Mean <dbl>, Jan_Median <dbl>,
#> #   Jan_Maximum <dbl>, Jan_Minimum <dbl>, Jan_P10 <dbl>, Jan_P20 <dbl>,
#> #   Feb_Mean <dbl>, Feb_Median <dbl>, Feb_Maximum <dbl>, Feb_Minimum <dbl>,
#> #   Feb_P10 <dbl>, Feb_P20 <dbl>, Mar_Mean <dbl>, Mar_Median <dbl>,
#> #   Mar_Maximum <dbl>, Mar_Minimum <dbl>, Mar_P10 <dbl>, Mar_P20 <dbl>,
#> #   Apr_Mean <dbl>, Apr_Median <dbl>, Apr_Maximum <dbl>, Apr_Minimum <dbl>,
#> #   Apr_P10 <dbl>, Apr_P20 <dbl>, May_Mean <dbl>, May_Median <dbl>,
#> #   May_Maximum <dbl>, May_Minimum <dbl>, May_P10 <dbl>, May_P20 <dbl>,
#> #   Jun_Mean <dbl>, Jun_Median <dbl>, Jun_Maximum <dbl>, Jun_Minimum <dbl>,
#> #   Jun_P10 <dbl>, Jun_P20 <dbl>, Jul_Mean <dbl>, Jul_Median <dbl>,
#> #   Jul_Maximum <dbl>, Jul_Minimum <dbl>, Jul_P10 <dbl>, Jul_P20 <dbl>,
#> #   Aug_Mean <dbl>, Aug_Median <dbl>, Aug_Maximum <dbl>, Aug_Minimum <dbl>,
#> #   Aug_P10 <dbl>, Aug_P20 <dbl>, Sep_Mean <dbl>, Sep_Median <dbl>,
#> #   Sep_Maximum <dbl>, Sep_Minimum <dbl>, Sep_P10 <dbl>, Sep_P20 <dbl>,
#> #   Oct_Mean <dbl>, Oct_Median <dbl>, Oct_Maximum <dbl>, Oct_Minimum <dbl>,
#> #   Oct_P10 <dbl>, Oct_P20 <dbl>, Nov_Mean <dbl>, Nov_Median <dbl>,
#> #   Nov_Maximum <dbl>, Nov_Minimum <dbl>, Nov_P10 <dbl>, Nov_P20 <dbl>,
#> #   Dec_Mean <dbl>, Dec_Median <dbl>, Dec_Maximum <dbl>, ...