Calculate long-term percentiles

Calculates the long-term percentiles from a daily streamflow data set. Calculates statistics from all values, unless specified. Returns a tibble with statistics.

Usage,

calc_longterm_percentile(
  data,
  dates = Date,
  values = Value,
  groups = STATION_NUMBER,
  station_number,
  percentiles,
  roll_days = 1,
  roll_align = "right",
  water_year_start = 1,
  start_year,
  end_year,
  exclude_years,
  complete_years = FALSE,
  months = 1:12,
  transpose = FALSE
)

Arguments

data

Data frame of daily data that contains columns of dates, flow values, and (optional) groups (e.g. station numbers). Leave blank if using station_number argument.

dates

Name of column in data that contains dates formatted YYYY-MM-DD. Only required if dates column name is not 'Date' (default). Leave blank if using station_number argument.

values

Name of column in data that contains numeric flow values, in units of cubic metres per second. Only required if values column name is not 'Value' (default). Leave blank if using station_number argument.

groups

Name of column in data that contains unique identifiers for different data sets, if applicable. Only required if groups column name is not 'STATION_NUMBER'. Function will automatically group by a column named 'STATION_NUMBER' if present. Remove the 'STATION_NUMBER' column beforehand to remove this grouping. Leave blank if using station_number argument.

station_number

Character string vector of seven digit Water Survey of Canada station numbers (e.g. "08NM116") of which to extract daily streamflow data from a HYDAT database. Requires tidyhydat package and a HYDAT database. Leave blank if using data argument.

percentiles

Numeric vector of percentiles (ex. c(5,10,25,75)) to calculate. Required.

roll_days

Numeric value of the number of days to apply a rolling mean. Default 1.

roll_align

Character string identifying the direction of the rolling mean from the specified date, either by the first ('left'), last ('right'), or middle ('center') day of the rolling n-day group of observations. Default 'right'.

water_year_start

Numeric value indicating the month (1 through 12) of the start of water year for analysis. Default 1.

start_year

Numeric value of the first year to consider for analysis. Leave blank to use the first year of the source data.

end_year

Numeric value of the last year to consider for analysis. Leave blank to use the last year of the source data.

exclude_years

Numeric vector of years to exclude from analysis. Leave blank to include all years.

complete_years

Logical values indicating whether to include only years with complete data in analysis. Default FALSE.

months

Numeric vector of months to include in analysis (e.g. 6:8 for Jun-Aug). Leave blank to summarize all months (default 1:12).

transpose

Logical value indicating whether to transpose rows and columns of results. Default FALSE.

Value

A tibble data frame of a long-term percentile of selected years and months.

Examples

# Run if HYDAT database has been downloaded (using tidyhydat::download_hydat())
if (file.exists(tidyhydat::hy_downloaded_db())) {

# Calculate the 20th percentile flow value from a flow record
calc_longterm_percentile(station_number = "08NM116",
                         percentile = 20)
                         
# Calculate the 90th percentile flow value with custom years
calc_longterm_percentile(station_number = "08NM116", 
                         start_year = 1980, 
                         end_year = 2010, 
                         percentile = 90)
                         
}
#> Warning: Calculation ignored missing values in data. Filter data to complete years or months if desired.
#> # A tibble: 1 x 2
#>   STATION_NUMBER   P90
#>   <chr>          <dbl>
#> 1 08NM116         19.3