Calculate annual summary and missing data statistics for screening data

Calculates means, medians, maximums, minimums, standard deviations of annual flows and data availability and missing data statistics for each year and month of each year. Calculates the statistics from all daily discharge values from all years, unless specified. Returns a tibble with statistics.

screen_flow_data(
  data,
  dates = Date,
  values = Value,
  groups = STATION_NUMBER,
  station_number,
  roll_days = 1,
  roll_align = "right",
  water_year_start = 1,
  start_year,
  end_year,
  months = 1:12,
  transpose = FALSE
)

Arguments

data
Data frame of daily data that contains columns of dates, flow values, and (optional) groups (e.g. station numbers). Leave blank if using station_number argument.

dates
Name of column in data that contains dates formatted YYYY-MM-DD. Only required if dates column name is not 'Date' (default). Leave blank if using station_number argument.

values
Name of column in data that contains numeric flow values, in units of cubic metres per second. Only required if values column name is not 'Value' (default). Leave blank if using station_number argument.

groups
Name of column in data that contains unique identifiers for different data sets, if applicable. Only required if groups column name is not 'STATION_NUMBER'. Function will automatically group by a column named 'STATION_NUMBER' if present. Remove the 'STATION_NUMBER' column beforehand to remove this grouping. Leave blank if using station_number argument.

station_number
Character string vector of seven digit Water Survey of Canada station numbers (e.g. "08NM116") of which to extract daily streamflow data from a HYDAT database. Requires tidyhydat package and a HYDAT database. Leave blank if using data argument.

roll_days
Numeric value of the number of days to apply a rolling mean. Default 1.

roll_align
Character string identifying the direction of the rolling mean from the specified date, either by the first ('left'), last ('right'), or middle ('center') day of the rolling n-day group of observations. Default 'right'.

water_year_start
Numeric value indicating the month (1 through 12) of the start of water year for analysis. Default 1.

start_year
Numeric value of the first year to consider for analysis. Leave blank to use the first year of the source data.

end_year
Numeric value of the last year to consider for analysis. Leave blank to use the last year of the source data.

months
Numeric vector of months to include in analysis (e.g. 6:8 for Jun-Aug). Leave blank to summarize all months (default 1:12).

transpose
Logical value indicating whether to transpose rows and columns of results. Default FALSE.

Value

A tibble data frame with the following columns:

Year

calendar or water year selected

n_days

number of days per year

n_Q

number of days per year with flow data

n_missing_Q

number of days per year with no flow data

Minimum

annual minimum of all daily flows for a given year

Maximum

annual maximum of all daily flows for a given year

Mean

annual mean of all daily flows for a given year

Median

annual median of all daily flows for a given year

StandardDeviation

annual 1 standard deviation of all daily flows for a given year

and the following monthly missing columns (order will depend on water_year_month):
Jan_missing_Q

number of Jan days per year with no flow data

Feb_missing_Q

number of Feb days per year with no flow data

Mar_missing_Q

number of Mar days per year with no flow data

Apr_missing_Q

number of Apr days per year with no flow data

May_missing_Q

number of May days per year with no flow data

Jun_missing_Q

number of Jun days per year with no flow data

Jul_missing_Q

number of Jul days per year with no flow data

Aug_missing_Q

number of Aug days per year with no flow data

Sep_missing_Q

number of Sep days per year with no flow data

Oct_missing_Q

number of Oct days per year with no flow data

Nov_missing_Q

number of Nov days per year with no flow data

Dec_missing_Q

number of Dec days per year with no flow data

Transposing data creates a column of "Statistics" and subsequent columns for each year selected.
Examples

# Run if HYDAT database has been downloaded (using tidyhydat::download_hydat())
if (file.exists(tidyhydat::hy_downloaded_db())) {

# Calculate screening statistics usinga data frame and data argument with defaults
flow_data <- tidyhydat::hy_daily_flows(station_number = "08NM116")
screen_flow_data(data = flow_data)

# Calculate screening statistics using station_number argument with defaults
screen_flow_data(station_number = "08NM116")
                  
# Calculate screening statistics for water years starting in October
screen_flow_data(station_number = "08NM116",
                 water_year_start = 9)
                  
# Calculate screening statistics for 7-day flows for July-September months only
screen_flow_data(station_number = "08NM116",
                 roll_days = 7,
                 months = 7:9)
                 
}

#> # A tibble: 69 x 13
#>    STATION_NUMBER  Year n_days   n_Q n_missing_Q Minimum Maximum  Mean Median
#>    <chr>          <dbl>  <int> <int>       <int>   <dbl>   <dbl> <dbl>  <dbl>
#>  1 08NM116         1949     92    92           0   0.838    4.29  1.52   1.32
#>  2 08NM116         1950     92    92           0   0.714   16.6   2.71   1.36
#>  3 08NM116         1951     92    92           0   0.629    8.80  2.21   1.23
#>  4 08NM116         1952     92    92           0   1.11    12.9   2.85   1.93
#>  5 08NM116         1953     92    92           0   0.821   24.1   4.77   2.24
#>  6 08NM116         1954     92    92           0   1.98    30.5   7.65   4.34
#>  7 08NM116         1955     92    92           0   0.396   19.9   4.82   2.35
#>  8 08NM116         1956     92    92           0   0.784    8.21  2.25   1.29
#>  9 08NM116         1957     92    92           0   0.916    6.23  2.36   1.92
#> 10 08NM116         1958     92    92           0   0.578    5.51  1.71   1.23
#> # ... with 59 more rows, and 4 more variables: StandardDeviation <dbl>,
#> #   Jul_missing_Q <int>, Aug_missing_Q <int>, Sep_missing_Q <int>